更新数据

+++ server/addons/supplier_search/model/entity/vendor_upload_file.go
@@ -15,8 +15,8 @@ type VendorUploadFile struct {
 	FileName            string      `json:"fileName"            description:"文件名称"`
 	FileId              int64       `json:"fileId"              description:"文件id"`
 	ExceptionDataFileId string      `json:"exceptionDataFileId" description:"异常数据文件id"`
-	ValidNum            int64       `json:"validNum"            description:"正常数据条数"`
-	ExceptionNum        int64       `json:"exceptionNum"        description:"异常数据条数"`
+	ValidNum            int         `json:"validNum"            description:"正常数据条数"`
+	ExceptionNum        int         `json:"exceptionNum"        description:"异常数据条数"`
 	AllColumn           string      `json:"allColumn"           description:"完整字段，多个用英文逗号隔开"`
 	CreatedAt           *gtime.Time `json:"createdAt"          description:"创建时间"`
 	UpdatedAt           *gtime.Time `json:"updatedAt"          description:"更新时间"`
